snmp.pdu_type: move unittests to tests/
